abstract out checklist
